+++ src/Parser/Context.php
@@ -10,21 +10,24 @@ use Xepozz\TestIt\Config;
 final class Context
 {
     public ?Node\Stmt\Namespace_ $namespace = null;
-    public ?Node\Stmt\Class_ $class = null;
+    public null|Node\Stmt\Class_|Node\Stmt\Enum_ $class = null;
     /**
-     * @var ?Node\Stmt\Class_[]
+     * @var Node\Stmt\Class_[]|Node\Stmt\Enum_[]
      */
     public array $classes = [];
     public ?Node\Stmt\ClassMethod $method = null;
+    public ?ClassContext $classContext = null;
+    public array $attributes = [];
 
     public function __construct(
         public readonly Config $config,
     ) {
     }
 
-    public function setClass(Node\Stmt\Class_ $node): void
+    public function setClass(Node\Stmt\Class_|Node\Stmt\Enum_ $node): void
     {
         $this->class = $node;
+        $this->classContext = new ClassContext();
         $this->classes[] = $node;
     }
 
@@ -36,6 +39,21 @@ final class Context
     public function setClassMethod(Node\Stmt\ClassMethod $node): void
     {
         $this->method = $node;
+    }
+
+    public function setAttribute(string $attribute, $value): void
+    {
+        $this->attributes[$attribute] = $value;
+    }
+
+    public function hasAttribute(string $attribute): bool
+    {
+        return isset($this->attributes[$attribute]);
+    }
+
+    public function getAttribute(string $attribute): mixed
+    {
+        return $this->attributes[$attribute];
     }
 
     public function reset(): void