+++ src/Parser/ContextMethodVisitor.php
@@ -11,9 +11,14 @@ use Nette\PhpGenerator\PhpNamespace;
 use PhpParser\Node;
 use PhpParser\NodeTraverser;
 use PhpParser\NodeVisitorAbstract;
+use Psr\EventDispatcher\EventDispatcherInterface;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\LoggerInterface;
+use Xepozz\TestIt\Event\AfterGenerationEvent;
+use Xepozz\TestIt\Event\ClassGeneratedEvent;
+use Xepozz\TestIt\Event\FileGeneratedEvent;
+use Xepozz\TestIt\Event\NamespaceGeneratedEvent;
 use Xepozz\TestIt\TestGenerator\ClassGenerator;
 use Xepozz\TestIt\TestGenerator\FileGenerator;
 use Xepozz\TestIt\TestGenerator\MethodGenerator;
@@ -43,6 +48,7 @@ final class ContextMethodVisitor extends NodeVisitorAbstract implements LoggerAw
         private readonly NamespaceGenerator $namespaceGenerator,
         private readonly ClassGenerator $classGenerator,
         private readonly MethodGenerator $methodGenerator,
+        private readonly EventDispatcherInterface $eventDispatcher,
     ) {
         $this->logger = $logger;
     }
@@ -59,9 +65,6 @@ final class ContextMethodVisitor extends NodeVisitorAbstract implements LoggerAw
 
     public function enterNode(Node $node): ?int
     {
-        if ($node instanceof Node\Stmt\Enum_) {
-            return NodeTraverser::DONT_TRAVERSE_CURRENT_AND_CHILDREN;
-        }
         if ($node instanceof Node\Stmt\Trait_) {
             return NodeTraverser::DONT_TRAVERSE_CURRENT_AND_CHILDREN;
         }
@@ -73,7 +76,7 @@ final class ContextMethodVisitor extends NodeVisitorAbstract implements LoggerAw
             $context->setNamespace($node);
             return null;
         }
-        if ($node instanceof Node\Stmt\Class_) {
+        if ($node instanceof Node\Stmt\Class_ || $node instanceof Node\Stmt\Enum_) {
             $this->generatedMethods = [];
             if ($this->isClassExcluded($node)) {
                 return NodeTraverser::DONT_TRAVERSE_CURRENT_AND_CHILDREN;
@@ -91,11 +94,8 @@ final class ContextMethodVisitor extends NodeVisitorAbstract implements LoggerAw
         return null;
     }
 
-    public function leaveNode(Node $node): null
+    public function leaveNode(Node $node)
     {
-        if ($node instanceof Node\Stmt\Enum_) {
-            return null;
-        }
         if ($node instanceof Node\Stmt\Trait_) {
             return null;
         }
@@ -104,16 +104,18 @@ final class ContextMethodVisitor extends NodeVisitorAbstract implements LoggerAw
         if ($node instanceof Node\Stmt\Namespace_) {
             $generated = $this->namespaceGenerator->generate($context, $this->generatedClasses);
             if ($generated !== null) {
+                $this->eventDispatcher->dispatch(new NamespaceGeneratedEvent($context, $generated));
                 $this->generatedNamespaces[] = $generated;
             }
             return null;
         }
-        if ($node instanceof Node\Stmt\Class_) {
+        if ($node instanceof Node\Stmt\Class_ || $node instanceof Node\Stmt\Enum_) {
             if ($this->isClassExcluded($node)) {
                 return null;
             }
             $generated = $this->classGenerator->generate($context, $this->generatedMethods);
             if ($generated !== null) {
+                $this->eventDispatcher->dispatch(new ClassGeneratedEvent($context, $generated));
                 $this->generatedClasses[] = $generated;
             }
             return null;
@@ -147,14 +149,22 @@ final class ContextMethodVisitor extends NodeVisitorAbstract implements LoggerAw
     {
         $files = [];
         foreach ($this->generatedNamespaces as $namespace) {
-            $files[] = $this->fileGenerator->generate([$namespace]);
+            $generated = $this->fileGenerator->generate([$namespace]);
+            $this->eventDispatcher->dispatch(new FileGeneratedEvent($this->contextProvider->getContext(), $generated));
+            $files[] = $generated;
         }
         return $files;
     }
 
-    private function isClassExcluded(Node\Stmt\Class_ $node): bool
+    public function afterTraverse(array $nodes)
     {
-        if ($node->isAbstract()) {
+        $this->eventDispatcher->dispatch(new AfterGenerationEvent($this->contextProvider->getContext()));
+        return parent::afterTraverse($nodes);
+    }
+
+    private function isClassExcluded(Node\Stmt\Class_|Node\Stmt\Enum_ $node): bool
+    {
+        if ($node instanceof Node\Stmt\Class_ && $node->isAbstract()) {
             return false;
         }
         $context = $this->contextProvider->getContext();