+++ src/TestGenerator/ClassGenerator.php
@@ -9,12 +9,10 @@ use Nette\PhpGenerator\Method;
 use PHPUnit\Framework\TestCase;
 use Xepozz\TestIt\Parser\Context;
 
-final readonly class ClassGenerator
+final class ClassGenerator
 {
     /**
-     * @param Context $context
      * @param Method[] $methods
-     * @return ClassType|null
      */
     public function generate(Context $context, array $methods): ?ClassType
     {