+++ src/TestMethodGenerator/NegativeMethodGenerator.php
@@ -16,18 +16,20 @@ use Xepozz\TestIt\PhpEntitiesConverter;
 use Xepozz\TestIt\TestGenerator\DataProviderGenerator;
 use Xepozz\TestIt\TypeNormalizer;
 use Xepozz\TestIt\TypeSerializer;
+use Xepozz\TestIt\ValueInitiator\ValueInitiatorInterface;
 
-final readonly class NegativeMethodGenerator implements TestMethodGeneratorInterface
+final class NegativeMethodGenerator implements TestMethodGeneratorInterface
 {
     public function __construct(
-        private TypeSerializer $typeSerializer,
-        private TypeNormalizer $typeNormalizer,
-        private MethodEvaluator $methodEvaluator,
-        private DataProviderGenerator $dataProviderGenerator,
-        private TestMethodFactory $testMethodFactory,
-        private PhpEntitiesConverter $phpEntitiesConverter,
-        private Dumper $dumper,
-        private MethodNameStrategy $methodNameStrategy,
+        private readonly TypeSerializer $typeSerializer,
+        private readonly TypeNormalizer $typeNormalizer,
+        private readonly MethodEvaluator $methodEvaluator,
+        private readonly DataProviderGenerator $dataProviderGenerator,
+        private readonly TestMethodFactory $testMethodFactory,
+        private readonly PhpEntitiesConverter $phpEntitiesConverter,
+        private readonly Dumper $dumper,
+        private readonly ValueInitiatorInterface $valueInitiator,
+        private readonly MethodNameStrategy $methodNameStrategy,
     ) {
     }
 
@@ -64,12 +66,18 @@ final readonly class NegativeMethodGenerator implements TestMethodGeneratorInter
                 ->setType($this->typeSerializer->serialize($this->typeNormalizer->denormalize($parameter->type)));
         }
         $arguments = $arguments === [] ? null : implode(', ', $arguments);
-        $variableName = '$' . lcfirst($class->name->name);
 
         $methodBodyBuilder = MethodBodyBuilder::create();
-        $methodBodyBuilder->addArrange("{$variableName} = new {$class->name->name}();");
-        $methodBodyBuilder->addAct("\$this->expectException(\$expectedExceptionClass);");
-        $methodBodyBuilder->addAct("{$variableName}->{$method->name->name}($arguments);");
+        if ($method->isStatic()) {
+            $methodBodyBuilder->addAct("\$this->expectException(\$expectedExceptionClass);");
+            $methodBodyBuilder->addAct("{$class->name->name}::{$method->name->name}($arguments);");
+        } else {
+            $variableName = '$' . lcfirst($class->name->name);
+            $classInitiation = $this->valueInitiator->getString($class);
+            $methodBodyBuilder->addArrange("{$variableName} = {$classInitiation};");
+            $methodBodyBuilder->addAct("\$this->expectException(\$expectedExceptionClass);");
+            $methodBodyBuilder->addAct("{$variableName}->{$method->name->name}($arguments);");
+        }
 
         $dataProviderNameParts = ['invalid', 'data', 'provider', $method->name->name];
         $dataProviderName = $this->methodNameStrategy->generate($context, $dataProviderNameParts);
@@ -106,7 +114,7 @@ final readonly class NegativeMethodGenerator implements TestMethodGeneratorInter
             $valuesToPrint = $this->phpEntitiesConverter->convert($case);
             try {
                 $this->methodEvaluator->evaluate($context, $valuesToPrint);
-            } catch (\Throwable $e) {
+            } catch (\Throwable) {
                 return true;
             }
         }