+++ src/ValueGenerator/ValueGeneratorRepository.php
@@ -4,8 +4,14 @@ declare(strict_types=1);
 
 namespace Xepozz\TestIt\ValueGenerator;
 
+use Xepozz\TestIt\Parser\ContextProvider;
+
 final class ValueGeneratorRepository
 {
+    public function __construct(
+        private readonly ContextProvider $contextProvider,
+    ) {
+    }
     public function getByType(mixed $possibleType): ?ValueGeneratorInterface
     {
         return match ($possibleType) {
@@ -16,6 +22,7 @@ final class ValueGeneratorRepository
             'int' => new IntegerValueGenerator(),
             'mixed' => new MixedValueGenerator(),
             'void' => new VoidValueGenerator(),
+            'self' => new ClassValueGenerator($this->contextProvider),
             default => null,
         };
     }