+++ tests/Integration/ConstructorParameters/src/EmptyParams.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Xepozz\TestIt\Tests\Integration\ConstructorParameters\src;
+
+class EmptyParams
+{
+    public function __construct()
+    {
+    }
+
+    public function inverse(bool $value): bool
+    {
+        return !$value;
+    }
+}